refactor queries to metric
