+++ hyrisecockpit/frontend/src/services/transformationService.ts
@@ -3,11 +3,12 @@ import {
   StorageData,
   TreemapDescription,
   AccessData,
-} from "../types/metrics";
+} from "@/types/metrics";
 import { TransformationService } from "@/types/services";
 import { useFormatting } from "@/meta/formatting";
 import { colorValueDefinition, multiColors } from "@/meta/colors";
 import { useDataEvents } from "@/meta/events";
+import { getDisplayedFromTransferred } from "@/meta/workloads";
 
 const { roundNumber, formatPercentage, formatTimeUnit } = useFormatting();
 const {
@@ -27,6 +28,7 @@ const transformationServiceMap: Record<Metric, TransformationService> = {
   ram: getRAMData,
   storage: getStorageData,
   throughput: getReadOnlyData,
+  queryInformation: getQueryInformationData,
 };
 
 /** export appropriate transformation function of given metric */
@@ -35,6 +37,19 @@ export function useDataTransformation(metric: Metric): TransformationService {
 }
 
 // TRANSFORM FUNCTION POOL
+
+/** transform query information data to table structure */
+function getQueryInformationData(data: any, primaryKey: string = ""): any {
+  const entry = data.find((dbEntry: any) => dbEntry.id === primaryKey);
+  return entry.query_information.map((query: any) => {
+    return {
+      queryNumber: query.query_number,
+      workloadType: getDisplayedFromTransferred(query.benchmark),
+      latency: roundNumber(query.latency, Math.pow(10, 6)),
+      throughput: query.throughput,
+    };
+  });
+}
 
 /** transform query type data to barchart structure */
 function getQueryTypeProportionData(data: any, primaryKey: string = ""): any {
@@ -405,7 +420,17 @@ export function useMaxValueHelper(
 ): ((data: any) => number) | undefined {
   const maxValueHelper: Partial<Record<Metric, (data: any) => number>> = {
     access: getAccessMaxValue,
+    queryInformation: getNumberOfQueries,
   };
+
+  /* detect max number of queries per db */
+  function getNumberOfQueries(data: any[]): number {
+    return data.reduce(
+      (numberOfQueries, queryData) =>
+        Math.max(numberOfQueries, queryData.query_information.length),
+      0
+    );
+  }
 
   /* detect max value of access data */
   function getAccessMaxValue(data: any): number {
