--- hyrisecockpit/frontend/src/types/queries.ts
@@ -1,13 +1,0 @@
-export type DetailedQueryInformation = {
-  queryNumber: string;
-  workloadType: string;
-  latency: number;
-  throughput: number;
-};
-
-export type DisplayedQueryInformation = {
-  queryNumber: string;
-  workloadType: string;
-  latency: string;
-  throughput: number;
-};
